+++ website/src/content/CHANGELOG.md
@@ -3,6 +3,15 @@
 ### Added
 
 - Added `Card` component.
+- Added a `2xs` size token with a value of `16rem` or `256px`.
+
+### Changed
+
+- Revised `Select` component to match the visual style of the `Input` component.
+
+### Fixed
+
+- Resolved an issue with the `Textarea` recipe where the resize handle would be hidden.
 
 ## [0.8.1] - 2023-08-17